Authjs - Autenticación y Protección de rutas
- Auth.js
- Proveedores
  * Google
  * Github
- Credenciales personalizadas
- Encriptación de contraseña
- Relaciones de base de datos
- Sesión de usuario
- Server y Client side validation
- Campos adicionales de usuario
- Modificación global de usuario de Auth.js

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from '@/auth/actions/auth-actions';
 import prisma from '@/lib/prisma';
 import { NextResponse, NextRequest } from 'next/server';
 import * as yup from 'yup';
@@ -31,11 +32,17 @@ const postSchema = yup.object({
   complete: yup.boolean().optional().default(false),
 });
 export async function POST(request: Request) {
+  const user = await getUserSessionServer();
+
+  if (!user) {
+    return NextResponse.json({ message: 'Usuario no autenticado' }, { status: 401 });
+  }
+
   try {
     const { complete, description } = await postSchema.validate(
       await request.json(),
     );
-    const todo = await prisma.todo.create({ data: { complete, description } });
+    const todo = await prisma.todo.create({ data: { complete, description, userId: user.id } });
 
     return NextResponse.json(todo);
   } catch (error) {
@@ -44,8 +51,14 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(request: Request) {
+  const user = await getUserSessionServer();
+
+  if (!user) {
+    return NextResponse.json({ message: 'Usuario no autenticado' }, { status: 401 });
+  }
+
   try {
-    await prisma.todo.deleteMany({ where: { complete: true } });
+    await prisma.todo.deleteMany({ where: { complete: true, userId: user.id } });
 
     return NextResponse.json('Borrado exitoso');
   } catch (error) {
